move premain to AgentBase
