+++ agent/src/main/java/eu/cqse/teamscale/jacoco/agent/AgentBase.java
@@ -1,9 +1,12 @@
 package eu.cqse.teamscale.jacoco.agent;
 
 import eu.cqse.teamscale.jacoco.agent.store.IXmlStore;
+import eu.cqse.teamscale.jacoco.util.LoggingUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.jacoco.agent.rt.RT;
+
+import java.lang.instrument.Instrumentation;
 
 /**
  * Base class for agent implementations. Handles logger shutdown,
@@ -33,6 +36,27 @@ public abstract class AgentBase {
 
 		logger.info("Starting JaCoCo agent with options: {}", options.getOriginalOptionsString());
 		logger.info("Storage method: {}", store.describe());
+	}
+
+	/** Called by the actual premain method once the agent is isolated from the rest of the application. */
+	public static void premain(String options, Instrumentation instrumentation) throws Exception {
+		AgentOptions agentOptions;
+		try {
+			agentOptions = AgentOptionsParser.parse(options);
+		} catch (AgentOptionParseException e) {
+			LoggingUtils.initializeDefaultLogging();
+			LogManager.getLogger(Agent.class).fatal("Failed to parse agent options: " + e.getMessage(), e);
+			System.err.println("Failed to parse agent options: " + e.getMessage());
+			throw e;
+		}
+
+		LoggingUtils.initializeLogging(agentOptions.getLoggingConfig());
+
+		LogManager.getLogger(Agent.class).info("Starting JaCoCo's agent");
+		org.jacoco.agent.rt.internal_c13123e.PreMain.premain(agentOptions.createJacocoAgentOptions(), instrumentation);
+
+		AgentBase agent = agentOptions.createAgent();
+		agent.registerShutdownHook();
 	}
 
 	/**
